source/data_engineering/airflow/requirements.txt
bs4
kafka-python-ng
lxml